src/processors/image_processor/requirements.txt
ocrmac==1.0.0; sys_platform == 'darwin'
pillow==11.1.0
pillow-heif==0.21.0